simple code

+++ dbsyncer-web/src/main/java/org/dbsyncer/web/controller/monitor/MonitorController.java
@@ -15,6 +15,7 @@ import org.dbsyncer.biz.enums.MetricEnum;
 import org.dbsyncer.biz.enums.StatisticEnum;
 import org.dbsyncer.biz.model.MetricResponse;
 import org.dbsyncer.biz.model.Sample;
+import org.dbsyncer.manager.impl.PreloadTemplate;
 import org.dbsyncer.web.controller.BaseController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -61,6 +62,9 @@ public class MonitorController extends BaseController {
     private SystemConfigService systemConfigService;
 
     @Resource
+    private PreloadTemplate preloadTemplate;
+
+    @Resource
     private MetricsEndpoint metricsEndpoint;
 
     @Resource
@@ -98,12 +102,16 @@ public class MonitorController extends BaseController {
 
     @Scheduled(fixedRate = 10000)
     public void refreshConnectorHealth() {
-        connectorService.refreshHealth();
+        if (preloadTemplate.isPreloadCompleted()) {
+            connectorService.refreshHealth();
+        }
     }
 
     @Scheduled(fixedRate = 30000)
     public void deleteExpiredDataAndLog() {
-        monitorService.deleteExpiredDataAndLog();
+        if (preloadTemplate.isPreloadCompleted()) {
+            monitorService.deleteExpiredDataAndLog();
+        }
     }
 
     @GetMapping("/queryData")
